% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGS_cbLDA.R
\name{CGS_clickbaitLDA}
\alias{CGS_clickbaitLDA}
\title{Click bait-LDA (CGS)}
\usage{
CGS_clickbaitLDA(
  w,
  x,
  alphastar,
  betaV,
  beta_ncb,
  beta_cb,
  b_doc,
  b_ncb,
  b_cb,
  TOPICS,
  vocab = NULL,
  iterations = 300,
  seed = 28,
  result_folder
)
}
\arguments{
\item{w}{A \eqn{D\times N_{\text{max}}} matrix \eqn{\mathbf{w}} of integers in \eqn{\{1,\ldots,V\}}.}

\item{x}{A \eqn{D^t}-dimensional vector \eqn{\bm{x}^t} of values in \eqn{\{0,1\}}, where \eqn{0leq D^t\leq D}.}

\item{alphastar}{A positive number \eqn{\alpha^*}.}

\item{betaV}{A positive number \eqn{\beta^V}.}

\item{beta_ncb}{A positive number \eqn{\beta^{ncb}}.}

\item{beta_cb}{A positive number \eqn{\beta^{cb}}.}

\item{b_ncb}{A \eqn{2}-dimensional vector \eqn{\bm{b}^{ncb}} of positive numbers.}

\item{b_cb}{A \eqn{2}-dimensional vector \eqn{\bm{b}^{cb}} of positive numbers.}

\item{TOPICS}{An integer number of topics \eqn{T}.}

\item{vocab}{A \eqn{V}-dimensional vector of vocabulary words.}

\item{iterations}{An integer number of iterations. Default is 300.}

\item{seed}{Seed. Default is 28.}

\item{result_folder}{A string specifying the folder in which the results will be saved.}

\item{bV}{A \eqn{2}-dimensional vector \eqn{\bm{b}^V} of positive numbers.}
}
\description{
Implementation of the Collapsed Gibbs Sampler (CGS) for click bait-LDA (cb-LDA).
}
\details{
Implementation in R and C++.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
