% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGS_HashtagLDA.R
\name{CGS_HashtagLDA}
\alias{CGS_HashtagLDA}
\title{Hashtag-LDA (CGS)}
\usage{
CGS_HashtagLDA(
  w,
  h,
  doc_users,
  alphastar,
  betaV,
  betaH,
  bH,
  iterations = 300,
  seed = 28,
  result_folder
)
}
\arguments{
\item{w}{A \eqn{D\times N_{\text{max}}} matrix \eqn{\mathbf{w}} of integers in \eqn{\{1,\ldots,V\}}.}

\item{h}{A \eqn{D\times L_{\text{max}}} matrix \eqn{\mathbf{h}} of integers in \eqn{\{1,\ldots,H\}}.}

\item{doc_users}{A \eqn{D}-dimensional vector of integers specifying the single author of each document. The number of authors is set to \code{U = max(doc_users)}.}

\item{alphastar}{A \eqn{T}-dimensional vector of positive numbers, \eqn{\bm{\alpha}^*}. The length of this vector specifies the number of topics \eqn{T}.}

\item{betaV}{A \eqn{V}-dimensional vector \eqn{\bm{\beta}^V} of positive numbers.}

\item{betaH}{A \eqn{H}-dimensional vector \eqn{\bm{\beta}^H} of positive numbers.}

\item{bH}{A \eqn{2}-dimensional vector \eqn{\bm{b}^H} of positive numbers.}

\item{iterations}{An integer number of iterations. Default is 300.}

\item{seed}{Seed. Default is 28.}

\item{result_folder}{A string specifying the folder in which the results will be saved.}
}
\description{
Implementation of the Collapsed Gibbs Sampler (CGS) for Hashtag-LDA.
}
\details{
Implementation in R and C++.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\seealso{
\code{\link{postproc_HashtagLDA}}, \code{\link{pred_HashtagLDA}}.
}
