% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGS_LDA.R
\name{CGS_LDA}
\alias{CGS_LDA}
\title{Latent Dirichlet Allocation (CGS)}
\usage{
CGS_LDA(w, alpha, betaV, iterations = 300, seed = 28, result_folder)
}
\arguments{
\item{w}{A \eqn{D\times N_{\text{max}}} matrix \eqn{\mathbf{w}} of integers in \eqn{\{1,\ldots,V\}}.}

\item{alpha}{A \eqn{T}-dimensional vector \eqn{\bm{\alpha}} of positive numbers. The length of this vector specifies the number of topics \eqn{T}.}

\item{betaV}{A \eqn{V}-dimensional vector \eqn{\bm{\beta}^V} of positive numbers.}

\item{iterations}{An integer number of iterations. Default is 300.}

\item{seed}{Seed. Default is 28.}

\item{result_folder}{A string specifying the folder in which the results will be saved.}
}
\description{
Implementation of the Collapsed Gibbs Sampler (CGS) for Latent Dirichlet Allocarion.
}
\details{
Implementation in R and C++.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\seealso{
\code{\link{postproc_LDA}}, \code{\link{pred_LDA}}.
}
