% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGS_MicroblogLDA.R
\name{CGS_MicroblogLDA}
\alias{CGS_MicroblogLDA}
\title{Microblog-LDA (CGS)}
\usage{
CGS_MicroblogLDA(
  w,
  doc_users,
  alphastar,
  alpha,
  beta,
  b,
  bdelta,
  bT,
  alpha0 = 10^-7,
  vocab = NULL,
  iterations = 300,
  seed = 28,
  result_folder
)
}
\arguments{
\item{w}{A list of \eqn{K} \eqn{D\times N_{\text{max}}^{(k)}} matrices \eqn{\mathbf{w}^{(k)}} of integers in \eqn{\{1,\ldots,V^{(k)}\}}.}

\item{doc_users}{A \eqn{D}-dimensional vector of integers specifying the single author of each document. The number of authors is set to \code{U = max(doc_users)}.}

\item{alphastar}{A \eqn{T}-dimensional vector \eqn{\bm{\alpha}^*} of positive numbers. The length of this vector specifies the number of topics \eqn{T}.}

\item{alpha}{A \eqn{T}-dimensional vector \eqn{\bm{\alpha}} of positive numbers. The length of this vector specifies the number of topics \eqn{T}.}

\item{beta}{A list of \eqn{K} \eqn{V^{(k)}}-dimensional vectors \eqn{\bm{\beta}^{(k)}} of positive numbers.}

\item{b}{A list of \eqn{K} \eqn{2}-dimensional vectors \eqn{\bm{b}^{(k)}} of positive numbers.}

\item{bdelta}{A \eqn{2}-dimensional vector \eqn{\bm{b}^\delta} of positive numbers.}

\item{bT}{A \eqn{2}-dimensional vector \eqn{\bm{b}^T} of positive numbers.}

\item{alpha0}{A positive number \eqn{\alpha_0}. Default is \eqn{10^{-7}}.}

\item{vocab}{A list of \eqn{V^{(k)}}-dimensional vectors of vocabulary words.}

\item{iterations}{An integer number of iterations. Default is 300.}

\item{seed}{Seed. Default is 28.}

\item{result_folder}{A string specifying the folder in which the results will be saved.}
}
\description{
Implementation of the Collapsed Gibbs Sampler (CGS) for Microblog-LDA.
}
\details{
Implementation in R and C++.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\seealso{
\code{\link{postproc_MicroblogLDA}}, \code{\link{pred_MicroblogLDA}}.
}
