% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGS_TwitterLDA.R
\name{CGS_TwitterLDA}
\alias{CGS_TwitterLDA}
\title{Twitter-LDA (CGS)}
\usage{
CGS_TwitterLDA(
  w,
  doc_users,
  alphastar,
  betaV,
  bV,
  vocab = NULL,
  iterations = 300,
  seed = 28,
  result_folder
)
}
\arguments{
\item{w}{A \eqn{D\times N_{\text{max}}} matrix \eqn{\mathbf{w}} of integers in \eqn{\{1,\ldots,V\}}.}

\item{doc_users}{A \eqn{D}-dimensional vector of integers specifying the single author of each document. The number of authors is set to \code{U = max(doc_users)}.}

\item{alphastar}{A \eqn{T}-dimensional vector \eqn{\bm{\alpha}^*} of positive numbers. The length of this vector specifies the number of topics \eqn{T}.}

\item{betaV}{A \eqn{V}-dimensional vector \eqn{\bm{\beta}^V} of positive numbers.}

\item{bV}{A \eqn{2}-dimensional vector \eqn{\bm{b}^V} of positive numbers.}

\item{vocab}{A \eqn{V}-dimensional vector of vocabulary words.}

\item{iterations}{An integer number of iterations. Default is 300.}

\item{seed}{Seed. Default is 28.}

\item{result_folder}{A string specifying the folder in which the results will be saved.}
}
\description{
Implementation of the Collapsed Gibbs Sampler (CGS) for Twitter-LDA.
}
\details{
Implementation in R and C++.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\seealso{
\code{\link{postproc_TwitterLDA}}, \code{\link{pred_TwitterLDA}}.
}
