% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dt_to_t}
\alias{dt_to_t}
\title{Document-term matrix to term matrix}
\usage{
dt_to_t(dt)
}
\arguments{
\item{dt}{A \eqn{D\times V} document-term matrix.}
}
\value{
A \eqn{D\times N_{\text{max}}} matrix of integers in \eqn{\{1,\ldots,V\}}.
}
\description{
Let \eqn{V} be the dimension of the vocabulary of the collection, this
function convert a  \eqn{D\times V} document-term matrix into a
\eqn{D\times N_{\text{max}}} matrix of integers in \eqn{\{1,\ldots,V\}}.
}
