% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_MLDA.R
\name{postproc_MicroblogLDA}
\alias{postproc_MicroblogLDA}
\title{Post-processing for Hashtag-LDA (CGS)}
\usage{
postproc_MicroblogLDA(
  result_folder,
  postproc_file,
  iterations = NULL,
  verbose = TRUE
)
}
\arguments{
\item{result_folder}{A string specifying the folder in which the results of a CGS obtained using the function \code{\link{CGS_MicroblogLDA}} are saved.}

\item{postproc_file}{A string specifying the RDS file in which the results will be saved. Write the name of the file without \code{.RDS} at the end.}

\item{iterations}{A vector of integers between 1 and the length of the chain which specifies the iterations to consider for the computation of the posterior means. Default is all the iterations.}

\item{verbose}{Logical: if \code{TRUE}, print the value of the logarithm of the a posteriori distribution at each iteration. Default is \code{TRUE}.}
}
\description{
Compute the logarithm of the a posteriori distribution at each MCMC iteration and the posterior means of the parameters.
}
\details{
Implementation in R and C++.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
\seealso{
\code{\link{CGS_MicroblogLDA}}, \code{\link{pred_MicroblogLDA}}.
}
