% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dt_to_t}
\alias{dt_to_t}
\title{Latent Dirichlet Allocation (CGS)}
\usage{
dt_to_t(dt, sparse = FALSE)
}
\arguments{
\item{dt}{A \eqn{D\times V} document-term matrix.}

\item{sparse}{Logical: if \code{TRUE}, the function returns a sparse matrix. Default is \code{FALSE}.}
}
\value{
A \eqn{D\times N_{\text{max}}} matrix of integers in \eqn{\{1,\ldots,V\}}.
}
\description{
Let \eqn{V} be the dimension of the vocabulary of the collection, this
function convert a  \eqn{D\times V} document-term matrix into a
\eqn{D\times N_{\text{max}}} matrix of integers in \eqn{\{1,\ldots,V\}}.
}
